requirements.txt
requests
beautifulsoup4
pillow
git+https://github.com/chrsbats/kvstore.git